licenses: move LicenseForm to Views/
Components::LicenseForm -> Views::Controllers::Licenses::Form
Per #4331. Two ERB callers updated. Also updates the LicenseForm
@example in application_form.rb's docstring to point at the new
location (the example was using the real class as illustration; now
illustrates the new pattern).
