+++ app/views/controllers/licenses/new.html.erb
@@ -3,4 +3,4 @@
   add_context_nav(license_form_new_tabs)
 %>
 
-<%= render(Components::LicenseForm.new(@license)) %>
+<%= render(Views::Controllers::Licenses::Form.new(@license)) %>
